Added ncvalues.h from vtknetcdf source to DataView to fix compilation on *nix.
This is a temporary bugfix for a VTK bug. The header is not installed by VTK
but is needed from the vtknetcdf_cxx library. On Windows this is not the
case because VTK is not installed but only built.

See http://www.vtk.org/pipermail/vtkusers/2012-January/120994.html for
more details.