Removed mandatory reference to a XML schema file in XML-IO classes.
These references are now hardcoded in the class constructors of the
XML-IO classes. Also the definition of the two search paths in
BaseLib::FileFinder from the main window are now moved into the
constructor of FileFinder.

XML schema files are first searched (for validation) by their path
and then by the reference in the XML file itself.