[MathLib] Constructor of TemplateVector3 needs a initializer list.
The TemplateVector3 is derived from TemplatePoint<T,N>. There isn't a
constructor taking 3 values for TemplatePoint. Instead of this a std::array
is used to initialize the TemplatePoint data structures. For this reason
the TemplateVector3 constructor that is taking two points to initialize
the internal data needs an initializer list.