[GL] OctTree::createOctTree returns pointer instead of obj.
At the moment it is not possible to write
oct_tree = createOctTree()
because the copy constructor is explicitely deleted. In my point of view
it does not make sense to implement a copy constructor. Sometimes it is
necessary to recreate the OctTree, for instance if the spatial size has changed.