[MeL/ME] Rm "error: unsigned integer overflow".
In the expressions "i+offset", "j+offset", ... the offset is converted to an 'unsigned int'.
If the value of the offset variable is negative this leads to a very large 'unsigned int' value.
The sum of i (j,...) and offset cannot be represented in type 'unsigned int'. This can be a problem.