[GL] lineSegmentIntersect: Add missing cases.
In that code setion it is already checked that the line segments (a,b) and (c,d)
are parallel. At this point it is also known that the line segment (a,c) is also
parallel to (a,b). In that case it is possible to express c as
c(t) = a + t * (b-a) (analog for the point d). Since the evaluation of all three
coordinate equations (x,y,z) have to lead to the same solution for the parameter
t it is sufficient to compute t only once.

Before the computation of the parameter was only performed for the x coordinate
equation. If the line segment are perpendicular to the x-axis this leads to
problems. The changes at hand treat this case.