[MeL] Return PropertyVector* instead of opt refs.
Since boost-1.61 (and also not in the std::optional)
optionals holding references do have some restrictions
and are difficult to use. In this special case
optional<PropertyVector&> is equivalent semantically
to a PropertyVector*. The pointer indicates that it is
not owning (contrary to usage of unique_ptr).

Thanks to Christoph pointing this issue out.