[MatL] Lubby2: Extract local temporal mat.props.
These values are only used inside the constitutive relation
computation and are not required to be saved for next
global iterations/time-steps.

Before this change the GM, eps_M_t, and eps_K_t values
were not initialized properly.

The loadState() function takes care of the latter two;
The values are extracted into own temporal structure.