[MaL] EigenMatrix: set/add values even if 0.
Under certain circumstances the sparsity pattern will be destroyed and
adding/setting values would then result in an expensive reallocation.

Assume that in the first global iteration only one half of the values is
set to non-zero. Before the linear solver is called the sparse matrix is
compressed thus reducing the allocated space in each row. In the
subsequent iterations, which might add or set the other half of the
matrix values reallocations would be necessary.