[App] Implement TestDefinition and test run.
Add --reference PATH option to ogs, which when present
will execute tests (vtkdiff) constructed from the
<test_definiton> section in the project file.
This section contains the tester (vtkdiff for now)
the file, field, and absolute and relative tolerances.

This allows the user to run tests w/o cmake and includes
the previously separated test description into the
project file.