[MGTL] Allow multiple bulk element ids in identify
If the lower dimensional elements are inside the bulk mesh,
there is no unique bulk element associated with it, but
multiple elements.
In the case of unique elements, the behaviour didn't change.
In the case of multiple elements, the bulk element ids are
stored in the FieldData with accompanying CellData array
'number_bulk_elements'.