[MatL] Major rewrite.
A lot of changes are renames for consistency with OGS.

Splitting files, and rename them (dropping mp, p, c prefixes).

Extraction of the create* functions lead to single step initialization
of the Medium, Phase, Component, and Property objects.

Formatting and clang-tidy.
Improve const-correctness.

Remove unneeded Undefined property: The nullptr is probably sufficient
in the PropertyArray, which is an array of pointers, for the property to
be undefined.

Unify value/dValue/d2Value access.

[App] Remove requirement for media ids sequence.

Reorder public, protected, private sections.

Add missing ogs_file_param documentation lines in the code.