[PaL] Add possibility to check parameter's domain.
Sometimes it happens, that a used parameter is defined
on a different mesh, then the mesh node/element ids do
not point to the right entities. This is difficult to
detect at runtime (you have to look into the results)
and leads to difficult to find errors.

This implements an additional function to check the
domains of definition, and adds an optional call to the
test when the parameter is found.