change the error message to warning
In this commit, the original OGS_FATAL function was changed to a WARN function. The reason behind is that, when running a HEAT_TRANSPORT_BHE process, the BHEs are sitting on a separate material ID. This means , the total number of material IDs is more than the number of media defined. By changing from OGS_FATAL to WARN, this allowing the process to be running.