[CL] Store cxxISolution in unique_ptr, not value.
This is motivated by deprecated (c++17) usage of std::ptr_fun
in phreeqc's Parser.h. The file is included in the chain
of includes and leaks the std::ptr_fun usage in the c++17
compiled code.
To break the dependency, the cxxISolution is forward-declared
and stored as a pointer instead of the value.