[CTest] Setup Python dependencies for benchmarks.
When a benchmark uses Python (REQUIREMENTS OGS_USE_PYTHON) and
there is a requirements.txt file in benchmarks project file
directory a Python virtual environment is created and dependencies
are installed during benchmark execution via ctest (inside the
build directory).

The PYTHONPATH environment variable of the benchmark test is set
to the virtual environment to mimic virtual environment activation.

Normally you would use

source .venv/bin/activate

but this is not possible when executing the benchmarks via ctest.