[CMake] Added version requirements file in json-format.
The file is read by CMake (via new a json parser submodule) and
the web site generator and provides info on required clang and
gcc versions. This info will be expanded in future commits.

Goal is to have a single source of truth for versions infos and
maybe other requirements.

On CMake you access the info via `ogs.`-prefixed variables:

    ${ogs.minimum_version.gcc}

In Hugo you access the info via the new dataFile-shortcode:

    {{< dataFile "versions.minimum_version.gcc" >}}