[ML] GL-integration for triangles with 7 IP
Added Gauss-Legendre Integration for Triangles with 7 Integration Points.
Necessary to output secondary variables at least in HM.
We might change the integration order to 5 as in the reference,
but I opted for 4 as it is the 4th option available.
Values are from here in Appendix II (p = 5):
Dunavant, D. A.
"High degree efficient symmetrical Gaussian quadrature rules for the triangle."
International journal for numerical methods in engineering
21.6 (1985): 1129-1148.
URL: https://onlinelibrary.wiley.com/doi/epdf/10.1002/nme.1620210612