[T/RM] Linear order shape functions test.
The reference files are converted from the second
order mesh results with the 'convertToLinearMesh'
utility.
Note, the sigma, and the NodalForces fields do not
compare well to the higher order results, but other
variables do very well.
sigma, probably due to extrapolation;
NodalForces because of the missing middle nodes which
would take some of the load.