[MPL] van Genuchten saturation uses p_b.
Choose neutral names for fitting parameters of the van Genuchten
capillary pressure-saturation model.

The original implementation was referring to "entry pressure", which is
not correct for name for this value.  The (van Genuchten 1980) paper
also does not refer to this parameter as "pressure" or "head" in any
way.

For details see also:

Jong van Lier Q, Pinheiro EAR. An alert regarding a common
misinterpretation of the Van Genuchten α parameter. Rev Bras Cienc Solo.
2018;42:e0170343.

accessed via
https://doi.org/10.1590/18069657rbcs20170343
http://www.scielo.br/pdf/rbcs/v42/0100-0683-rbcs-42-e0170343.pdf