[PL/Output] Extend findPVDFile.
Since for one process output on multiple
(sub)domains / meshes can be required it is
necessary to manage multiple pvd files. All pvd
files for all processes are stored in a multimap.
The findPVDFile method is extended to find the
correct pvd file for the corresponding domain and
process.

For processes containing multiple subprocesses
(staggered scheme and component transport) it
isn't allowed to contain the process id in the
output prefix template.