[PL] Fix bug in TimeLoop::outputSolutions.
For the staggered coupling some additional work has
to be done in the same scope of the remaining code.

The call to preTimestep uses internally Process::_coupled_solutions
which is a pointer to an object of type
CoupledSolutionsForStaggeredScheme.
This variable is set by setCoupledSolutionsForStaggeredScheme()
using the address of a *local* variable that will
be destroyed when the execution of the code is out
of the scope. For this reason the call to
preTimestep have to be inside the same scope.