[MPL] Function-type Property using exprtk.
A generalized property type for (almost) arbitrary functions. Based on
the exprtk library also used in the Function-type parameters.

In this initial implementation all *scalar* variables from the variable
array passed for the value and derivative evaluation can be used. Return
values are either scalar, 2- or 3-vectors, and 2x2- or 3x3 matrices.

The evaluation is somewhat slower than the dedicated implementation; in
the case of component transport process a bilinear property was replaced
by exactly the same function-type property and the total run-time for
the ctest has increased by around 6%.