[A/U/ME] Add switch arg to tool.
A part of an element is inside of a polygon if at
least one node is in the (convex) polygon. Sometimes
it is useful to forster that all nodes of an
element have to be in the polygon. With the new
implemented switch argument the user can choose the
behaviour of the algorithm.