[GL/MGTL] Cleanup Pointer to vector of Point's.
Pointer to vector of Point's wasn't cleaned up properly.

In rotatePolygonToXY a pointer to the vector of
polygon points was created. The rotated polygon is
on this vector. When the polygon is returned from
the function it isn't possible to cleanup the vector
properly. So the creation of the polygon is move to
the function markNodesOutSideOfPolygon.