[MeL/VTK] Safely convert all (u)int8-64 bits types
Adding conversion of short = int16 types.

Merging conversion of vtkLongArray with vtkLongLongArray, and of
vtkUnsignedLongArray with vtkUnsignedLongLongArray.
For some systems the conversion of vtkLongArray, where the size of long
and long long types are different is not available. (But could be added
to the int type array conversion.)

Actually the use of vtkLongArray is discouraged by VTK: "The C++
standard does not define the exact size of the long type, so use of this
type directly is discouraged. If an array of 32 bit integers is needed,
prefer vtkTypeInt32Array to this class. If an array of 64 bit integers
is needed, prefer vtkTypeInt64Array to this class." -- from vtkLongArray
doxygen documentation.

Add compile-time tests for conversion of integer types since in OGS we
use the char, short, int, long types which are not guarantied to be of
corresponding length of vtk types.