[PL/THM] Comment incomplete implementation
The implementation of the compressibility effects in the temperature
equation is not complete, especially the MTu and KTp terms are missing
in the rhs_T.
Having this terms in the Jacobian only does not change the results but
breaks the linearization.

When thermal osmosis is off, the now commented rhs_T term leads to
nonphysical temperature rise.