[MPL] Add early exit in if saturation change is 0
This change is twofold; one one side the complicated
matrix multiplication is avoided, but on the other side
it is also preventing in initialization of the processes,
when dt = 0 the division by zero.

Naturally, this is a hack due to lag of separation of
constitutive variables update in the time step from
initialization (where dt, S_L_prev relevant for this
property, are not yet initialized).