[MeL] Add move-ctor to Mesh, disable assignments
Saves few deep copies in the current implementation of
the deactivated subdomains and can be further used
to eliminate the unique_ptr return type pattern
occuring in the construction of final or intermediate
objects.

Explicitly delete Mesh assignments:
Because the mesh id is unique (const member set from global
counter) the assignments should be forbidden (explicitly)
because comparison of meshes by ids would be broken.