[PL/BC] Reimplement Robin-type BC
The former implementation did not produce expected results for 2D case
(and probably for 3D case). Rewriting and implementing the Jacobian for
the Newton scheme. Keeping K matrix for the Picard scheme as applying
the BC to the r.h.s. only significantly increases number of non-linear
iterations because the problem becomes non-linear.