[T] 2D Robin BC tests; Newton and Picard
A method of manufactured solutions is used to verify implementation of
the Robin-type boundary condition. Grid convergence was assessed and the
observed order of convergence is 1.99, which, as expected, is close
enough to 2.

In the same setup a Neumann and Dirichlet-type boundary conditions can
(and were) tested with same convergence order as the Robin-type BC.

The test is testing both non-linear solver schemes, Newton and Picard.
With the new implementation of the Robin-type BC the convergence rates
of the non-linear solver are satisfying, usually 2 iterations per time
step for both schemes.