[MeL] Use exact predicates for point projection
Previously using inexact predicates would yield ambiguous results whether a point
is inside or outside of a projected element, depending on compiler and architecture.

The exact predicates are slower, so one could think about to add a run-time flag to
use exact or inexact predicates giving the user an option for fast or correct calculations.