[PL] Use localDOF() in THM
In one place an evaluation was pulled into the integration point loop.
That might cause a tiny bit more computation. On the other hand the
optimizer might take care of that. On the plus side we don't map data by
hand anymore in these places, and the evaluations are more local than
before.

localDOF() seems to be incompatible with the current approach for
staggered coupling. But I have the impression that there are some flaws
with the current approach to staggered coupling.