[T/SD] Excavation replace reference
On domain deactivation, and especially on reactivation (backfilling)
the current implementation is/was using the previously stored
value for the strain, which is not entirely correct.
Much better would be to set the secondary variables to
a given value upon deactivation/reactivation. But this
is out of scope for now.