[ogs] -m switch: fallback to reading gml from prj-file directory.
When a mesh directory is given with the -m switch, the gml file is also
read from the given mesh directory. This commit adds a fallback to also
reading from the prj-file directory.

Motivation is when using partitioned meshes they often are created in
subdirectories of the project. Before this one had to copy the gml file
into that location too.